app/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-jwt-extended==4.5.3
flask-cors==4.0.0
flask-migrate==4.0.5
werkzeug==2.3.7
python-dotenv==1.0.0
bcrypt==4.0.1
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
textblob==0.17.1
requests==2.31.0